+++ oxe-api/resource/public/get_public_companies.py
@@ -48,7 +48,7 @@ class GetPublicCompanies(MethodResource, Resource):
         if "count" in kwargs and kwargs["count"] is True:
             response = {"count": self.db.get_filtered_companies(kwargs).count()}
         else:
-            entities = c.id, c.name, c.legal_status, c.is_startup, c.is_cybersecurity_core_business, \
+            entities = c.id, c.name, c.headline, c.legal_status, c.is_startup, c.is_cybersecurity_core_business, \
                 c.trade_register_number, c.creation_date, c.description, c.website, c.image, c.status, c.linkedin_url, \
                 c.twitter_url, c.youtube_url, c.discord_url, c.sync_node, c.sync_id, c.sync_global, c.sync_address, \
                 c.sync_status