+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Network graph tab on Dashboard page on oxe-web-admin
 - resource user/get_user_company_assignments
 - Value "SELECT" on FormQuestion type
+- "headline" field for entities
 
 ### Changed